requirements.txt
attrs==20.3.0
cycler==0.10.0
flake8==3.8.4
flake8-docstrings==1.5.0
iniconfig==1.1.1
joblib==1.0.1
kiwisolver==1.3.1
kornia @ git+https://github.com/kornia/kornia@08e5a309739d7ab99728d5573cf448219f05c36d
llvmlite==0.35.0
matplotlib==3.3.4
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numba==0.52.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pynndescent==0.5.2
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
snowballstemmer==2.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
umap-learn==0.5.1
